pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sf2hs"
version = "0.1.0"
description = "A tool for automating Salesforce to HubSpot metadata synchronization"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "simple-salesforce>=1.12.0",
    "hubspot-api-client>=7.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pandas>=2.0.3",
    "openpyxl>=3.1.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
sf2hs = "sf2hs.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/sf2hs"]

[tool.black]
line-length = 120
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 120
target-version = "py38"
